+++ lib/ui/groups/GroupMembersPanel.dart
@@ -415,19 +415,21 @@ class _GroupMembersPanelState extends State<GroupMembersPanel> implements Notifi
   }
 
   Widget _buildStatusDismissLayer() {
-    return Positioned.fill(child:
-      BlockSemantics(child:
-        GestureDetector(
-          onTap: () {
-            Analytics().logSelect(target: 'Close Dropdown');
-            setState(() {
-              _statusValuesVisible = false;
-            });
-          },
-          child: Container(color: Styles().colors!.blackTransparent06)
-        )
-      )
-    );
+    return
+    Container(
+        constraints: BoxConstraints(minHeight:  MediaQuery.of(context).size.height),
+        child: BlockSemantics(child:
+            GestureDetector(
+              onTap: () {
+                Analytics().logSelect(target: 'Close Dropdown');
+                setState(() {
+                  _statusValuesVisible = false;
+                });
+              },
+              child: Container(color: Styles().colors!.blackTransparent06)
+            )
+          )
+        );
   }
 
   Widget _buildStatusValuesWidget() {