+++ lib/ui/home/HomeInboxWidget.dart
@@ -19,8 +19,9 @@ import 'package:rokwire_plugin/service/connectivity.dart';
 import 'package:rokwire_plugin/service/inbox.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
-import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
+
+enum HomeInboxContent { all, unread }
 
 class HomeInboxWidget extends StatefulWidget {
 
@@ -63,6 +64,10 @@ class _HomeInboxWidgetState extends State<HomeInboxWidget> implements Notificati
   Key _pageViewKey = UniqueKey();
   Map<String, GlobalKey> _contentKeys = <String, GlobalKey>{};
   final double _pageSpacing = 16;
+
+  static const String localScheme = 'local';
+  static const String allNotificationsHost = 'all_notifications';
+  static const String localUrlMacro = '{{local_url}}';
 
   @override
   void initState() {
@@ -212,9 +217,19 @@ class _HomeInboxWidgetState extends State<HomeInboxWidget> implements Notificati
     else if (_loadingMessages) {
       return HomeProgressWidget();
     }
+    else if (_messages == null) {
+      return HomeMessageCard(message: Localization().getStringEx("widget.home.inbox.text.failed.description", "Failed to load notifications."),);
+    }
     else if (CollectionUtils.isEmpty(_messages)) {
-      return HomeMessageCard(
-        message: Localization().getStringEx("widget.home.inbox.text.empty.description", "Notifications are available right now."),);
+      if (widget.content == HomeInboxContent.unread) {
+        String message = Localization().getStringEx("widget.home.inbox.text.empty.unread.description", "You have no unread notifications. <a href='$localUrlMacro'><b>View all notifications</b></а>.")
+          .replaceAll(localUrlMacro, '$localScheme://$allNotificationsHost');
+          return HomeMessageHtmlCard(message: message, onTapLink: _onTapMessageLink,);
+      }
+      else {
+        return HomeMessageCard(
+          message: Localization().getStringEx("widget.home.inbox.text.empty.all.description", "You have not any notifications yet."),);
+      }
     }
     else {
       return _buildMessagesContent();
@@ -230,18 +245,6 @@ class _HomeInboxWidgetState extends State<HomeInboxWidget> implements Notificati
       for (InboxMessage message in _messages!) {
         pages.add(Padding(key: _contentKeys[message.messageId ?? ''] ??= GlobalKey(), padding: EdgeInsets.only(right: _pageSpacing, bottom: 16), child:
           InboxMessageCard(message: message, onTap: () => _onTapMessage(message)),
-        ));
-      }
-
-      if (_loadingMessagesPage) {
-        pages.add(Padding(key: _contentKeys['last'] ??= GlobalKey(), padding: EdgeInsets.only(right: _pageSpacing), child:
-          Container(decoration: BoxDecoration(color: Styles().colors?.white, borderRadius: BorderRadius.circular(5)), child:
-            HomeProgressWidget(
-              padding: EdgeInsets.symmetric(horizontal: 32, vertical: (_pageHeight - 24) / 2),
-              progessSize: Size(24, 24),
-              progressColor: Styles().colors?.fillColorPrimary,
-            ),
-          ),
         ));
       }
 
@@ -298,6 +301,15 @@ class _HomeInboxWidgetState extends State<HomeInboxWidget> implements Notificati
     return minContentHeight ?? 0;
   }
 
+  void _onTapMessageLink(String? url) {
+    Uri? uri = (url != null) ? Uri.tryParse(url) : null;
+    if (uri?.scheme == localScheme) {
+      if (uri?.host.toLowerCase() == allNotificationsHost.toLowerCase()) {
+        SettingsNotificationsContentPanel.present(context, content: SettingsNotificationsContent.all);
+      }
+    }
+  }
+
   void _onTapMessage(InboxMessage message) {
     Analytics().logSelect(target: message.subject);
     SettingsNotificationsContentPanel.launchMessageDetail(message);
@@ -308,5 +320,3 @@ class _HomeInboxWidgetState extends State<HomeInboxWidget> implements Notificati
     SettingsNotificationsContentPanel.present(context, content: (_unread == true) ? SettingsNotificationsContent.unread : SettingsNotificationsContent.all);
   }
 }
-
-enum HomeInboxContent { all, unread }